Move Binary Ninja helpers to dedicated package
