+++ sc62015/pysc62015/asm.py
@@ -1,5 +1,5 @@
 from typing import Any, List, Optional, Union, cast, TypedDict, Type, Literal
-from . import binja_api  # noqa: F401
+from binja_helpers import binja_api  # noqa: F401
 from binaryninja.architecture import RegisterName
 from lark import Lark, Transformer, Token
 from .instr import (
