+++ sc62015/pysc62015/emulator.py
@@ -10,7 +10,7 @@ from .instr import (
     OPCODES,
     Instruction,
 )
-from .mock_llil import (
+from binja_helpers.mock_llil import (
     MockLowLevelILFunction,
     MockLLIL,
     MockLabel,
